Drop motion odom header
